+++ packages/api/src/agent.ts
@@ -95,6 +95,7 @@ export class AtpAgent {
         handle: res.data.handle,
         did: res.data.did,
         email: opts.email,
+        emailConfirmed: false,
       }
       return res
     } catch (e) {
@@ -126,6 +127,7 @@ export class AtpAgent {
         handle: res.data.handle,
         did: res.data.did,
         email: res.data.email,
+        emailConfirmed: res.data.emailConfirmed,
       }
       return res
     } catch (e) {
@@ -154,6 +156,7 @@ export class AtpAgent {
       }
       this.session.email = res.data.email
       this.session.handle = res.data.handle
+      this.session.emailConfirmed = res.data.emailConfirmed
       return res
     } catch (e) {
       this.session = undefined
@@ -268,6 +271,7 @@ export class AtpAgent {
     } else if (isNewSessionObject(this._baseClient, res.body)) {
       // succeeded, update the session
       this.session = {
+        ...(this.session || {}),
         accessJwt: res.body.accessJwt,
         refreshJwt: res.body.refreshJwt,
         handle: res.body.handle,