+++ packages/dev-env/src/bin.ts
@@ -1,5 +1,6 @@
 import { generateMockSetup } from './mock'
 import { TestNetwork } from './network'
+import { mockMailer } from './util'
 
 const run = async () => {
   console.log(`
@@ -23,6 +24,7 @@ const run = async () => {
     },
     plc: { port: 2582 },
   })
+  mockMailer(network.pds)
   await generateMockSetup(network)
 
   console.log(