+++ packages/pds/src/db/database-schema.ts
@@ -15,6 +15,7 @@ import * as notification from './tables/user-notification'
 import * as blob from './tables/blob'
 import * as repoBlob from './tables/repo-blob'
 import * as deleteAccountToken from './tables/delete-account-token'
+import * as emailToken from './tables/email-token'
 import * as moderation from './tables/moderation'
 import * as mute from './tables/mute'
 import * as listMute from './tables/list-mute'
@@ -40,6 +41,7 @@ export type DatabaseSchemaType = runtimeFlag.PartialDB &
   blob.PartialDB &
   repoBlob.PartialDB &
   deleteAccountToken.PartialDB &
+  emailToken.PartialDB &
   moderation.PartialDB &
   mute.PartialDB &
   listMute.PartialDB &