+++ packages/aws/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atproto/aws
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atproto/repo@0.3.2
+
 ## 0.1.1
 
 ### Patch Changes