+++ packages/repo/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @atproto/repo
 
+## 0.3.2
+
+### Patch Changes
+
+- Updated dependencies [[`41ee177f`](https://github.com/bluesky-social/atproto/commit/41ee177f5a440490280d17acd8a89bcddaffb23b)]:
+  - @atproto/common-web@0.2.1
+  - @atproto/common@0.3.1
+  - @atproto/identity@0.2.1
+  - @atproto/lexicon@0.2.2
+  - @atproto/syntax@0.1.2
+
 ## 0.3.1
 
 ### Patch Changes