+++ packages/xrpc-server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atproto/xrpc-server
 
+## 0.3.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atproto/common@0.3.1
+  - @atproto/lexicon@0.2.2
+
 ## 0.3.1
 
 ### Patch Changes